==> step05_new_features.ipynb <==
# Шаг 5. Создание новых признаков

### Импорт необходимых библиотек

import pandas as pd

### Импорт датафрейма

df = pd.read_csv('data/df_after_step03.csv', index_col=0)
df.head()

### Добавление новых признаков

Помимо признаков, добавленных на втором шаге (`first_workday`, `second_workday` - были добавлены на раннем этапе, т.к. строки с выходными днями, необходимые для их создания, были преимущественно удалены, поскольку в них было много пропусков).

Первый признак, который можно добавить - волатильность (`imoex_vol`), а точнее изменение значения индекса Мосбиржи за день. По смыслу это добавит компонент с настроением на финансовом рынке.

df['imoex_vol'] = df['imoex_close'] - df['imoex_open']

Также можно попробовать добавить индикатор мягкости или жесткости ДКП в России (`ru_monetary`) и США (`us_monetary`). Несомненно, это грубая аппроксимация, но вариант, как это можно сделать: 
    
    1) вычесть из ИПЦ 100 (чтобы получить инфляцию в процентах)
    2) умножить ИПЦ на 12, чтобы получить инфляцию за год ("грубость" метода лежит тут; но это вынужденная мера, т.к. ставка ЦБ / ФРС - годовая)
    3) вычесть из ставки ЦБ / ФРС полученную "инфляцию"

df['ru_monetary'] = df['cb_key_rate'] - (df['ru_cpi'] - 100) * 12
df['us_monetary'] = df['fed_rate'] - (df['us_cpi'] - 100) * 12

df.head()

df.to_csv('data/df_after_step05.csv')
