==> step07_ml.ipynb <==
# Шаг 7. Машинное обучение

### Импорт необходимых библиотек

import pandas as pd

# функции из sklearn импортируются далее по файлу в начале каждого раздела

df = pd.read_csv('df_after_step05.csv', index_col=0)

## Что предсказываем и почему
USD/RUB

_work in progress_

## Предварительная подготовка данных

Для того, чтобы обучать модели, нужно:

- Преобразовать нечисловые признаки: перевести все категориальные признаки в числовой формат
- Разделить выборки на обучающую (train) и тестовую (test)
- Отмасштабировать признаки

### Преобразование нечисловых признаков

# поскольку категориальная переменная бинарна, можно не увеличивать размерность и оставить всё в одной колонке
df_ml = pd.get_dummies(df, drop_first=True)

# так как название задваивается, вернем его к нормальному виду
df_ml.rename(columns={'workday_workday': 'workday'}, inplace=True)

df_ml.head()

### Разделение выборки на тестовую и обучающую

from sklearn.model_selection import train_test_split

y = df_ml['usdrub']
X = df_ml.drop(columns=['usdrub'])

X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=7, test_size=0.2)

### Масштабирование признаков

Выберем, при помощи чего масштабировать данные:
- **StandardScaler**: 
    - Плюсы: учитывает и сохраняет специфику данных
    - Минусы: предполагает нормальное распределение. Как можно увидеть на графиках в [step04](https://github.com/evelinakolomoets/andan_project/blob/main/step04_plots.ipynb), наши данные не распределены по нормальному закону
    - **Вывод**: скорее не подходит
- **MinMaxScaler**:
    - Плюсы: интуитивно понятно
    - Минусы: может потеряться информация, связанная со спецификой данных
    - **Вывод**: скорее не подходит

- **RobustScaler**:
    - Плюсы: хорошо работает с асимметрией в данных (как видно на [графиках](https://github.com/evelinakolomoets/andan_project/blob/main/step04_plots.ipynb), она присутствует в нашем датасете)
    - Минусы: никак не учитываются среднее и медиана
    - **Вывод**: подходит лучше, чем первые два варианта


from sklearn.preprocessing import RobustScaler

scaler = RobustScaler()

X_train_norm = scaler.fit_transform(X_train)
X_test_norm = scaler.transform(X_test)
